backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pymongo
passlib==1.7.4
bcrypt==4.0.1
python-jose
email-validator
pandas
numpy
joblib
xgboost
scikit-learnpytest
pytest-cov